Cập nhật module Indoor: tự động xóa nguyên liệu quá hạn, tối ưu lại thống báo nếu lỗi,thêm chức năng xem công thức nấu ăn, đè xuất món ăn có thể nấu bằng nguyên liệu sẵn có trong tủ lạnh

+++ src/main/java/org/example/repository/indoor/RecipeIngredientRepository.java
@@ -6,6 +6,6 @@ import org.springframework.stereotype.Repository;
 import java.util.List;
 
 @Repository
-public interface RecipeIngredientRepository extends JpaRepository<RecipeIngredient, Long> {
-    List<RecipeIngredient> findByRecipeId(Long recipeId);
+public interface RecipeIngredientRepository extends JpaRepository<RecipeIngredient, Integer> {
+    List<RecipeIngredient> findByRecipeId(Integer recipeId);
 }
